requirements.txt
altgraph==0.17.4
blis==0.7.11
cachetools==4.2.4
catalogue==2.0.10
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
comtypes==1.3.1
confection==0.1.4
contextvars==2.4
cymem==2.0.8
dataclasses==0.8
down==1.0
Flask==2.0.3
idna==3.6
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
langcodes==3.3.0
lxml==5.1.0
MarkupSafe==2.0.1
murmurhash==1.0.10
numpy==1.19.5
packaging==21.3
pathy==0.10.3
pefile==2023.2.7
preshed==3.0.9
psutil @ file:///C:/ci/psutil_1612298125479/work
pychrome==0.2.4
pydantic==1.9.2
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
PyJWT==2.4.0
pyparsing==3.1.2
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-docx==0.8.11
pywin32==305
pywin32-ctypes==0.2.2
pywinauto==0.6.8
requests==2.27.1
selenium==3.141.0
six==1.16.0
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==0.0.33
srsly==2.4.8
tencentcloud-sdk-python==3.0.1109
thinc==8.1.12
tqdm==4.64.1
typer==0.9.0
typing_extensions==4.1.1
urllib3==1.26.18
wasabi==1.1.2
websocket-client==1.3.1
Werkzeug==2.0.3
wincertstore==0.2
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.6.0/zh_core_web_sm-3.6.0-py3-none-any.whl
zhipuai==1.0.7
zipp==3.6.0
